Optimize low level structure operations
Three versions of revert, playing with reservation of vectors, rewritten get_used_symbols, added get_max_symbol, push_back and back to Post, Move, OrdVector ...

backup

reserve_on_insert

some minor changes

some minor changes

Playing some more, optimizing some more. Simple revert is still wining, fragile revert is close and still has some potential in optimizing the internal data structure (one could make make it a simple flat array).
Strange how much take push_back takes everywhere.

removing a mess with renaming of revert

removing some useless changes

removing errors in ubuntu compilation

reorganizing Util

some small things

better get_useful_states

trim2 works!

still something weird

fix of NumberPredicate::defragment

rewrite of get_max_state of delta, fix of move in delta::shake_down (don't move to yourself)

remove nonsense

stopped working ...

Now it works. I was a hell, I am not a programmer, I want to grow potatoes.

comment

+++ include/mata/alphabet.hh
@@ -23,7 +23,8 @@
 #include <mata/ord-vector.hh>
 
 namespace Mata {
-        using Symbol = unsigned long;
+        //using Symbol = unsigned long;
+        using Symbol = unsigned;
         using StringToSymbolMap = std::unordered_map<std::string, Symbol>;
 
          /**
